% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isOutside.R
\name{isOutside}
\alias{isOutside}
\title{Is the vector in the extension?}
\usage{
isOutside(x)
}
\arguments{
\item{x}{Vector that is tested}
}
\value{
Returns true if any of the entries is negative
}
\description{
Is the vector in the extension?
}
