% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_bounds.R
\name{knapsack_bounds}
\alias{knapsack_bounds}
\title{Find the smallest bounds of the knapsack}
\usage{
knapsack_bounds(A, CPLB_idx, objective.in, y)
}
\arguments{
\item{A}{Configuration matrix}

\item{CPLB_idx}{A2 index}

\item{objective.in}{Vector a of definition of the knapsack}

\item{y}{Vector of observations}
}
\value{
Vector a which defines the bounds of the hypercuboid
}
\description{
Find the smallest bounds of the knapsack
}
