% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_sampler.R
\name{knapsack_sampler}
\alias{knapsack_sampler}
\title{Sampler over the knapsack-extension}
\usage{
knapsack_sampler(
  A,
  y,
  n.samples = 1e+05,
  n.chains = 4,
  n.burnin = 10000,
  B = NULL,
  CPLB_idx = NULL,
  x.start = NULL,
  a = 0,
  b = 0,
  ldelta = 0
)
}
\arguments{
\item{A}{Configuration matrix}

\item{y}{Observation vector}

\item{n.samples}{Number of samples per chain}

\item{n.chains}{Number of chains}

\item{n.burnin}{length of burnin}

\item{B}{A set of vectors connecting the M_p-extension}

\item{CPLB_idx}{A set indicating A_2 columns}

\item{x.start}{Initial point}

\item{a}{Objective function of the knapsack set description}

\item{b}{Upper bound of the knapsack set describtion}

\item{ldelta}{log(delta) determining epsilon}
}
\value{
Returns a matrix containing the samples over the M_p-extension generated by a hit-and-run algorithm
}
\description{
Sampler over the knapsack-extension
}
