% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_proportion.R
\name{positive_proportion}
\alias{positive_proportion}
\title{Calculate the proportion of positive samples}
\usage{
positive_proportion(output)
}
\arguments{
\item{output}{Matrix containing the sample including chain, iteration and draw numbers.}
}
\value{
Returns Returns the proportion of positive samples in the extension sample
}
\description{
Calculate the proportion of positive samples
}
