% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposePoint_hypercube.R
\name{proposePoint_hypercube}
\alias{proposePoint_hypercube}
\title{Propose a new point in the hypercube extension}
\usage{
proposePoint_hypercube(x, B, CPLB_idx, a)
}
\arguments{
\item{x}{Current point}

\item{B}{Sampling directions (Should be a column partition lattice basis)}

\item{CPLB_idx}{Index of the A_2 matrix}

\item{a}{Bounds of the hypercuboid}
}
\value{
Returns a new proposed point in the hypercube extension
}
\description{
Propose a new point in the hypercube extension
}
