% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposePoint_knapsack.R
\name{proposePoint_knapsack}
\alias{proposePoint_knapsack}
\title{Propose new point in the knapsack extension}
\usage{
proposePoint_knapsack(x, B, CPLB_idx, a, b)
}
\arguments{
\item{x}{Current point}

\item{B}{Sampling directions (Should be a CPLB)}

\item{CPLB_idx}{Index of columns of matrix for A_2}

\item{a}{Objective function of the knapsack set definition}

\item{b}{Upper limit of the knapsack set definition}
}
\value{
Returns a vector containing the new proposed point in the knapsack extension
}
\description{
Propose new point in the knapsack extension
}
