% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposePoint_p.R
\name{proposePoint_p}
\alias{proposePoint_p}
\title{Proposes a new point in the M_p-extension}
\usage{
proposePoint_p(x, B, p)
}
\arguments{
\item{x}{Current point}

\item{B}{Matrix of sampling directions}

\item{p}{Relaxation limit}
}
\value{
Returns a new proposed vector in the M_p-extension
}
\description{
Proposes a new point in the M_p-extension
}
