% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_negative_rows.R
\name{remove_negative_rows}
\alias{remove_negative_rows}
\title{Removing negative rows of the output matrix}
\usage{
remove_negative_rows(output)
}
\arguments{
\item{output}{Matrix containing the sample including chain, iteration and draw numbers.}
}
\value{
Returns sample matrix without any negative vectors.
}
\description{
Removing negative rows of the output matrix
}
