% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_draws_df.R
\name{to_draws_df}
\alias{to_draws_df}
\title{Converting from matrix to draws_df}
\usage{
to_draws_df(output, var.names = NULL)
}
\arguments{
\item{output}{Matrix containing the sample including chain, iteration and draw numbers.}

\item{var.names}{Fill names for the variables if NULL names will be $xi$ for the ith variable.}
}
\value{
Returns the sample object of any of the samplers as a draws_df file
}
\description{
Converting from matrix to draws_df
}
