% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypercube_bounds.R
\name{hypercube_bounds}
\alias{hypercube_bounds}
\title{Find the smallest bounds of the hypercube}
\usage{
hypercube_bounds(A, CPLB_idx, y)
}
\arguments{
\item{A}{Configuration matrix}

\item{CPLB_idx}{A2 index}

\item{y}{Vector of observations}
}
\value{
Vector a which defines the bounds of the hypercuboid
}
\description{
Find the smallest bounds of the hypercube
}
